requirements.txt
# Core ML/Audio
librosa>=0.9.0
openl3>=0.4.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0

# APIs and Web
spotipy>=2.20.0
flask>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0

# Audio Processing
soundfile>=0.10.0

# Development
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pytest>=6.0.0